lesson 8 update FileManager
